+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.10.0] - 2022-12-24
+### Added
+- Added .NET 7 support
+
+### Changed
+- Update `Microsoft.NETFramework.ReferenceAssemblies` to v1.0.3
+
+### Removed
+- Removed .NET Core 3.1 (LTS) support
+- Removed ctor `ShamirsSecretSharing(IExtendedGcdAlgorithm<TNumber> extendedGcd, int securityLevel)`.
+- Removed method `MakeShares(TNumber numberOfMinimumShares, TNumber numberOfShares)`.
+
 ## [0.9.0] - 2022-10-09
 ### Added
 - Add `ToInt32()` method to `BigIntCalculator` and `Calculator` class.
@@ -166,7 +178,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `LICENSE.md`
 - Added `README.md`
 
-[Unreleased]: https://github.com/shinji-san/SecretSharingDotNet/compare/v0.9.0...HEAD
+[0.10.0]: https://github.com/shinji-san/SecretSharingDotNet/compare/v0.9.0...0.10.0
 [0.9.0]: https://github.com/shinji-san/SecretSharingDotNet/compare/v0.8.0...v0.9.0
 [0.8.0]: https://github.com/shinji-san/SecretSharingDotNet/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/shinji-san/SecretSharingDotNet/compare/v0.6.0...v0.7.0